Start of summary functionality
