+++ src/views/project/publishPartition/PublishPartitionView.vue
@@ -17,28 +17,7 @@
         - check if success of submitting request
             - IF YES : your request to publish the {name} partition in P('projectId') has been submitted. You will receive an email notification at ('email') when duplication of the partition is completed.
         	- if no: Error copying partition: (error) // maybe keep
-    4. Summary (dear lord...)
-        - print partition name
-        - display partition contents
-            - just all the nums of what was retrieved
-        - display to the user that if they wish to edit any of this then they must do so in the matrix editor
-        - if onetime media is present
-            - inform the user: This partition contains {vn_num_one_time_use_media} media released for onetime use. Due to this license, these media may not be usable in a derivative project. You must either change the media reuse license, or choose to keep the media in the current project, the new parition based project or both (bypassing one-time use).  
-                - Click here to see a list of all onetime use media.
-                - display three options 
-                    - Keep one-time use media in existing project
-                    - Move one-time use media to new partition based project
-                    - Use one-time use media in both existing project and new partition based project
-        - print the help texts
-            - When you publish this partition, a new project will be created with the characters and taxa listed above. Any matrices, media, copyright documents and bibliographic
-			  citations associated with the characters and taxa will be copied into the new project. The new project will be left in an unpublished state to allow you time to review
-			  prior to publication.
-            - Note that the new project you create here will be separate from P<?php print $t_project->getPrimaryKey(); ?> (<?php print $t_project->get('name'); ?>) and create
-			  only the subset of information described above. Changes made to the newly created project will not affect P<?php print $t_project->getPrimaryKey(); ?> or vice versa.
-            - Note that publishing a partition may consume significant server resources, particularly storage. Please be considerate and only publish
-			  when you need to.
-        - print if partition is empty: You cannot publish this partition because it is empty</strong>. Please add characters and taxa and try again.
-        - else show a button that allows publishment
+
 
     
     Setup layout
@@ -51,15 +30,6 @@
         - get partition info to print
         - get whether it was successful or not 
         - try to tell user what went wrong with the request
-    4. Summary (dear lord...)
-        - get taxa and characters from the partition 
-        - get num of matricies and get media specimens and views 
-        - get num of media, onetime media, specimens, and views
-            - make it possible to get links to each onetimeMedia when requested
-            - make it possible to retrieve the decision with what has to be done for the onetimeMedia
-        - get num of documents, labels, and bib references
-        - create new taxa?
-        - check if the partition is empty 
 */
 import axios from 'axios'
 import router from '@/router'
@@ -69,9 +39,6 @@ import { useRoute } from 'vue-router'
 const route = useRoute()
 const projectId = route.params.id
 let partitions = ref([])
-let onetimeMedia = null
-let taxa = null
-let character = null
 
 onMounted(() => {
   getPartitions()
@@ -86,9 +53,14 @@ async function getPartitions() {
   partitions.value = response.data
 }
 
-async function publishPartition() {
+function selectPartition(event) {
   const formData = new FormData(event.target)
   const formObject = Object.fromEntries(formData.entries())
+
+  const partitionId = formObject.selectedPartition
+  router.push({
+    path: `/myprojects/${projectId}/publish/partition/${partitionId}`,
+  })
 }
 </script>
 
@@ -120,7 +92,7 @@ async function publishPartition() {
   </p>
 
   <div v-if="partitions.length > 0">
-    <form @submit.prevent="publishPartition">
+    <form @submit.prevent="selectPartition">
       <p>Please select a partition to publish:</p>
       <ul class="list-group">
         <div class="list-group-item-content">
@@ -133,7 +105,7 @@ async function publishPartition() {
               class="form-check-input"
               type="radio"
               name="selectedPartition"
-              value="partition.partition_id"
+              :value="partition.partition_id"
             />
             <div class="list-group-item-name">
               {{ partition.name }}
@@ -141,9 +113,10 @@ async function publishPartition() {
           </li>
         </div>
       </ul>
+      <button type="submit">Start Publishing Process</button>
     </form>
   </div>
   <div v-else>
-    <p>{{ partitions }}</p>
+    <p>Currently, there are no partitions associated with this project.</p>
   </div>
 </template>
